requirements.txt
torch == 2.6.0
torchvision == 0.21.0
numpy == 2.2.3
matplotlib == 3.10.1
tqdm == 4.62.3
shap == 0.46.0
pillow ==11.1.0